Add basic JSDoc
