+++ war/src/main/js/pages/project/filter-build-history.types.js
@@ -1,0 +1,17 @@
+/**
+ * @typedef QueryParameters
+ * @type {object}
+ * @property {string | undefined} search
+ * @property {string | undefined} older-than
+ * @property {string | undefined} newer-than
+ */
+
+/**
+ * @typedef CardControlsOptions
+ * @type {object}
+ * @property {boolean} banana
+ * @property {boolean} page-has-up
+ * @property {boolean} page-has-down
+ * @property {string | undefined} page-entry-newest
+ * @property {string | undefined} page-entry-oldest
+ */
