+++ war/webpack.config.js
@@ -52,8 +52,8 @@ module.exports = (env, argv) => ({
     "components/row-selection-controller": [
       path.join(__dirname, "src/main/js/components/row-selection-controller"),
     ],
-    "filter-build-history": [
-      path.join(__dirname, "src/main/js/filter-build-history.js"),
+    "pages/project/filter-build-history": [
+      path.join(__dirname, "src/main/js/pages/project/filter-build-history.js"),
     ],
     "simple-page": [path.join(__dirname, "src/main/scss/simple-page.scss")],
     styles: [path.join(__dirname, "src/main/scss/styles.scss")],
